+++ ToDoList/ToDoListApp/views/edit_task_view.py
@@ -9,23 +9,24 @@ from ToDoListApp.models import Task, ToDoList
 @login_required
 def edit_task(request, task_id, list_id):
     task = Task.objects.get(pk=task_id)
-    if request.method == 'POST':
+    if request.method == "POST":
         form = TaskForm(request.POST)
         if form.is_valid():
-            task.title = form.cleaned_data['title']
-            task.description = form.cleaned_data['description']
-            task.deadline = form.cleaned_data['deadline']
-            task.priority = form.cleaned_data['priority']
-            try:
-                file = request.FILES['attachment']
+            task.title = form.cleaned_data["title"]
+            task.description = form.cleaned_data["description"]
+            task.deadline = form.cleaned_data["deadline"]
+            task.priority = form.cleaned_data["priority"]
+
+            if hasattr(request, "FILES") and "attachment" in request.FILES:
+                file = request.FILES["attachment"]
                 task.attachment = file
-            except:
-                pass
-            try:
-                if form.data['attachment-clear'] == 'on':
-                    task.attachment.delete()
-            except:
-                pass
+
+            if (
+                "attachment-clear" in form.data
+                and form.data["attachment-clear"] == "on"
+            ):
+                task.attachment.delete()
+
             task.save()
             messages.add_message(request, messages.INFO, "Task edited successfully.")
 
@@ -35,14 +36,23 @@ def edit_task(request, task_id, list_id):
                 messages.add_message(request, messages.ERROR, error[1])
 
         to_do_list = ToDoList.objects.get(pk=list_id)
-        sorted_tasks = to_do_list.tasks.all().order_by('deadline', 'priority')
-        return render(request, "v1/get_list_template.html",
-                      {"tasks": sorted_tasks, "to_do_list": to_do_list,
-                       "user": to_do_list.owner})
+        sorted_tasks = to_do_list.tasks.all().order_by("deadline", "priority")
+        return render(
+            request,
+            "v1/get_list_template.html",
+            {"tasks": sorted_tasks, "to_do_list": to_do_list, "user": to_do_list.owner},
+        )
 
     else:
-        form = TaskForm(initial={'title': task.title, 'description': task.description,
-                                 'deadline': task.deadline, 'priority': task.priority,
-                                 'attachment': task.attachment})
-        return render(request, "v1/edit_task_template.html",
-                      {"task": task, "form": form})
+        form = TaskForm(
+            initial={
+                "title": task.title,
+                "description": task.description,
+                "deadline": task.deadline,
+                "priority": task.priority,
+                "attachment": task.attachment,
+            }
+        )
+        return render(
+            request, "v1/edit_task_template.html", {"task": task, "form": form}
+        )
